requirements.txt
streamlit>=1.35.0
plotly>=5.20.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0
google-generativeai>=0.8.0
scipy>=1.12.0
reportlab>=4.0.0
matplotlib>=3.8.0
